# myHDFW-frontend-vite/src/api/api-docs.yaml
{ "openapi": "3.0.1","info": { "title": "myHDFW REST API","description": "","version": "1.0" },"servers": [ { "url": "http://localhost:8080/api/v1","description": "Generated server url" } ],"security": [ { "Bearer Authentication": [ ] } ],"paths": { "/room": { "get": { "tags": [ "room-controller" ],"operationId": "getAllRooms","responses": { "200": { "description": "Found Rooms","content": { "*/*": { "schema": { "type": "string" } } } } } },"post": { "tags": [ "room-controller" ],"operationId": "createRoom","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Room" } } },"required": true },"responses": { "201": { "description": "Room created","content": { "*/*": { "schema": { "$ref": "#/components/schemas/Room" } } } } } } },"/auth/login": { "post": { "tags": [ "auth-controller" ],"operationId": "login","requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LoginRequest" } } },"required": true },"responses": { "200": { "description": "Login Successful","content": { "*/*": { "schema": { "type": "string" } } } } } } },"/room/{room_id}": { "get": { "tags": [ "room-controller" ],"operationId": "getRoom","parameters": [ { "name": "room_id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "200": { "description": "Found Room","content": { "*/*": { "schema": { "$ref": "#/components/schemas/Room" } } } } } },"delete": { "tags": [ "room-controller" ],"operationId": "deleteRoom","parameters": [ { "name": "room_id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "204": { "description": "Room deleted" } } } },"/location": { "get": { "tags": [ "location-controller" ],"operationId": "getAllLocations","responses": { "200": { "description": "Found Locations","content": { "*/*": { "schema": { "type": "string" } } } } } } },"/location/{location_id}": { "get": { "tags": [ "location-controller" ],"operationId": "getLocation","parameters": [ { "name": "location_id","in": "path","required": true,"schema": { "type": "integer","format": "int64" } } ],"responses": { "200": { "description": "Found Location","content": { "*/*": { "schema": { "$ref": "#/components/schemas/Location" } } } } } } } },"components": { "schemas": { "Location": { "type": "object","properties": { "id": { "type": "integer","format": "int64" },"name": { "type": "string" },"rooms": { "uniqueItems": true,"type": "array","items": { "$ref": "#/components/schemas/Room" } } } },"Room": { "type": "object","properties": { "id": { "type": "integer","format": "int64" },"name": { "type": "string" },"capacity": { "type": "integer","format": "int32" },"location": { "$ref": "#/components/schemas/Location" } } },"LoginRequest": { "type": "object","properties": { "username": { "type": "string" },"password": { "type": "string" } } } },"securitySchemes": { "Bearer Authentication": { "type": "http","scheme": "bearer","bearerFormat": "JWT" } } } }